+++ element/al_ico_element.js
@@ -1,0 +1,17 @@
+/**
+ * The al-ico element
+ *
+ * @author   Jelle De Loecker   <jelle@elevenways.be>
+ * @since    0.6.0
+ * @version  0.6.0
+ */
+const Icon = Function.inherits('Alchemy.Element', 'AlIco');
+
+/**
+ * The stylesheet to load for this element
+ *
+ * @author   Jelle De Loecker <jelle@elevenways.be>
+ * @since    0.6.0
+ * @version  0.6.0
+ */
+Icon.setStylesheetFile('alchemy_icons');